requirements.txt
streamlit
spotipy
python-dotenv
google-generativeai
pandas
